+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.4 - **NOT RELEASED**
     - Add support for Go C Strings with tests.
+    - Skip `003-multiple.t` on macOS since Go runtime always crashes.
 
 ## 0.0.3 - 1 Dec 2017
     - Add change log.